+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Head
 
 - Fix handling of `^` (exponentiation) and `%` (remainder) expression operators.
+- Fix handling of empty input to `formulaToExpression`. It now returns `undefined`.
 
 ## 0.2.0